+++ core/src/main/java/org/pilotpi/api/IMUListener.java
@@ -1,5 +1,5 @@
 package org.pilotpi.api;
 
 public interface IMUListener {
-	void update(float roll, float pitch, float yaw);
+	void update(double roll, double pitch, double yaw);
 }
